requirements.txt
streamlit
pandas
prophet
google-genai
plotly
